requirements.txt
crispy-bootstrap5==0.7
Django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-htmx==1.13.0
django_debug_toolbar==3.8.1
django-croppie==1.3
Pillow==9.4.0
kolo==2.3.0